# [Backend] cuda normalize and permute, cuda concat, optimized ppcls, ppdet & ppseg