--- a/fastdeploy/core/fd_tensor.h
+++ b/fastdeploy/core/fd_tensor.h
@@ -39,7 +39,9 @@ struct FASTDEPLOY_DECL FDTensor {
   // GPU to inference the model
   // so we can skip data transfer, which may improve the efficience
   Device device = Device::CPU;
-  int device_id = 0;
+  // By default the device id of FDTensor is -1, which means this value is
+  // invalid, and FDTensor is using the same device id as Runtime.
+  int device_id = -1;
 
   // Whether the data buffer is in pinned memory, which is allocated
   // with cudaMallocHost()