--- a/fastdeploy/runtime.cc
+++ b/fastdeploy/runtime.cc
@@ -568,6 +568,10 @@ std::vector<TensorInfo> Runtime::GetOutputInfos() {
 
 bool Runtime::Infer(std::vector<FDTensor>& input_tensors,
                     std::vector<FDTensor>* output_tensors) {
+  for (auto& tensor: input_tensors) {
+    FDASSERT(FDtensor.device_id == option.device_id,
+             "Input tensor and runtime have different device ids.");
+  }
   return backend_->Infer(input_tensors, output_tensors);
 }