--- a/fastdeploy/vision/classification/ppcls/model.cc
+++ b/fastdeploy/vision/classification/ppcls/model.cc
@@ -23,7 +23,7 @@ PaddleClasModel::PaddleClasModel(const std::string& model_file,
                                  const std::string& config_file,
                                  const RuntimeOption& custom_option,
                                  const ModelFormat& model_format)
-    : preprocessor_(config_file, (custom_option.device == Device::GPU)) {
+    : preprocessor_(config_file) {
   valid_cpu_backends = {Backend::ORT, Backend::OPENVINO, Backend::PDINFER,
                         Backend::LITE};
   valid_gpu_backends = {Backend::ORT, Backend::PDINFER, Backend::TRT};