--- a/fastdeploy/vision/classification/ppcls/preprocessor.cc
+++ b/fastdeploy/vision/classification/ppcls/preprocessor.cc
@@ -20,12 +20,15 @@ namespace fastdeploy {
 namespace vision {
 namespace classification {
 
-PaddleClasPreprocessor::PaddleClasPreprocessor(const std::string& config_file) {
-  FDASSERT(BuildPreprocessPipelineFromConfig(config_file), "Failed to create PaddleClasPreprocessor.");
+PaddleClasPreprocessor::PaddleClasPreprocessor(const std::string& config_file,
+                                               bool use_cuda) {
+  FDASSERT(BuildPreprocessPipelineFromConfig(config_file, use_cuda),
+           "Failed to create PaddleClasPreprocessor.");
   initialized_ = true;
 }
 
-bool PaddleClasPreprocessor::BuildPreprocessPipelineFromConfig(const std::string& config_file) {
+bool PaddleClasPreprocessor::BuildPreprocessPipelineFromConfig(
+    const std::string& config_file, bool use_cuda) {
   processors_.clear();
   YAML::Node cfg;
   try {
@@ -70,6 +73,12 @@ bool PaddleClasPreprocessor::BuildPreprocessPipelineFromConfig(const std::string
 
   // Fusion will improve performance
   FuseTransforms(&processors_);
+
+  // Set all processors to use CUDA, but only those support CUDA can actually
+  // use CUDA
+  for (size_t i = 0; use_cuda && i < processors_.size(); ++i) {
+    processors_[i]->UseCuda();
+  }
   return true;
 }