--- a/fastdeploy/vision/classification/ppcls/preprocessor.h
+++ b/fastdeploy/vision/classification/ppcls/preprocessor.h
@@ -28,8 +28,7 @@ class FASTDEPLOY_DECL PaddleClasPreprocessor {
    *
    * \param[in] config_file Path of configuration file for deployment, e.g resnet/infer_cfg.yml
    */
-  explicit PaddleClasPreprocessor(const std::string& config_file,
-                                  bool use_cuda = false);
+  explicit PaddleClasPreprocessor(const std::string& config_file);
 
   /** \brief Process the input image and prepare input tensors for runtime
    *
@@ -41,10 +40,10 @@ class FASTDEPLOY_DECL PaddleClasPreprocessor {
 
 
  private:
-  bool BuildPreprocessPipelineFromConfig(const std::string& config_file,
-                                         bool use_cuda = false);
+  bool BuildPreprocessPipelineFromConfig(const std::string& config_file);
   std::vector<std::shared_ptr<Processor>> processors_;
   bool initialized_ = false;
+  bool use_cuda_ = false;
 };
 
 }  // namespace classification