--- a/fastdeploy/vision/common/processors/base.h
+++ b/fastdeploy/vision/common/processors/base.h
@@ -58,11 +58,16 @@ class FASTDEPLOY_DECL Processor {
 
   virtual bool operator()(Mat* mat, ProcLib lib = ProcLib::DEFAULT);
 
+  void UseCuda() {
+    use_cuda_ = true;
+  }
+
  protected:
   FDTensor* UpdateAndGetReusedBuffer(
       const std::vector<int64_t>& new_shape, const int& opencv_dtype,
       const std::string& buffer_name, const Device& new_device = Device::CPU,
       const bool& use_pinned_memory = false);
+  bool use_cuda_ = false;
 
  private:
   std::unordered_map<std::string, FDTensor> reused_buffers_;