--- a/fastdeploy/vision/detection/ppdet/ppyoloe.cc
+++ b/fastdeploy/vision/detection/ppdet/ppyoloe.cc
@@ -147,6 +147,13 @@ bool PPYOLOE::BuildPreprocessPipelineFromConfig() {
   // Fusion will improve performance
   FuseTransforms(&processors_);
 
+  // Set all processors to use CUDA, but only those support CUDA can actually
+  // use CUDA
+  if (runtime_option.device == Device::GPU) {
+    for (size_t i = 0; i < processors_.size(); ++i) {
+      processors_[i]->UseCuda();
+    }
+  }
   return true;
 }
 
@@ -160,8 +167,6 @@ bool PPYOLOE::Preprocess(Mat* mat, std::vector<FDTensor>* outputs) {
       return false;
     }
   }
-
-  Cast::Run(mat, "float");
 
   outputs->resize(2);
   (*outputs)[0].name = InputInfoOfRuntime(0).name;