--- a/fastdeploy/vision/detection/ppdet/ppyoloe.cc
+++ b/fastdeploy/vision/detection/ppdet/ppyoloe.cc
@@ -146,14 +146,6 @@ bool PPYOLOE::BuildPreprocessPipelineFromConfig() {
 
   // Fusion will improve performance
   FuseTransforms(&processors_);
-
-  // Set all processors to use CUDA, but only those support CUDA can actually
-  // use CUDA
-  if (runtime_option.device == Device::GPU) {
-    for (size_t i = 0; i < processors_.size(); ++i) {
-      processors_[i]->UseCuda();
-    }
-  }
   return true;
 }