--- a/python/fastdeploy/vision/classification/ppcls/__init__.py
+++ b/python/fastdeploy/vision/classification/ppcls/__init__.py
@@ -35,10 +35,12 @@ class PaddleClasPreprocessor:
         """
         return self._preprocessor.run(input_ims)
 
-    def use_cuda(self):
+    def use_gpu(self, gpu_id=0):
         """Use CUDA preprocessors
+
+        :param: gpu_id: GPU device id
         """
-        return self._preprocessor.use_cuda()
+        return self._preprocessor.use_gpu(gpu_id)
 
 
 class PaddleClasPostprocessor: