<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types(PR类型)
<!-- One of [ Model | Backend | Serving | Quantization | Doc | Bug Fix | Other] -->

Backend

### Describe
<!-- Describe what this PR does -->

- CUDA kernel for Color convert + Normalize + Permute
- Concat support GPU
- Applied this optimization to PPClas, PPDet and PPSeg
- Tested on PPLCNetV2, PPYOLOE and UNet.