@wang-xinyu 2022-11-10
PPClas End2end test on T4, TRT8.4
Latency in ms.

Version | Model | FP32 | FP16 | INT8
-- | -- | -- | -- | --
0.3.0 | PP-LCNetv2 | 2.23 | 1.87 | 1.78
This | PP-LCNetv2 | 1.54 | 1.03 | 0.89