@wang-xinyu 2022-11-10
PPDet End2end test on T4, TRT8.4
Latency in ms.

.   | Model | FP32
-- | -- | --
Before | PPYOLOE | 35.05
After | PPYOLOE | 32.97

Inference is 32.6ms, so Preprocessing is 2.45ms -> 0.37ms