@wang-xinyu 2022-11-10
PPSeg UNet  PPInfer GPU backend, T4.

Latency in ms.

 | Input size | Preprocess latency, before | after |
 | -- | -- | -- | 
 | 2048x1024 | 33.921 | 1.405 |
 | 320x160 | 0.228 | 0.194 |