@wang-xinyu 2022-11-14
> 另外预处理设置的gpu_id，如果与runtime的gpu_id不符，还需处理ORT、Paddle、和TensorRT在接收到输入后的情况

@jiangjiajun @heliqi  我们是在三个backend里面加device id的判断和cudaMemcpyPeerAsync()吗？   还是device id不一致的时候Assert？      如果不给用户提示就直接copy到不同的device，可能会影响性能