@heliqi 2022-11-14
> > 另外预处理设置的gpu_id，如果与runtime的gpu_id不符，还需处理ORT、Paddle、和TensorRT在接收到输入后的情况
> 
> @jiangjiajun @heliqi 我们是在三个backend里面加device id的判断和cudaMemcpyPeerAsync()吗？ 还是device id不一致的时候Assert？ 如果不给用户提示就直接copy到不同的device，可能会影响性能

我建议先Assert掉，至少给个警告。 
1. 正常使用Model这种简单接口时不推荐用户把预处理和runtime设置在不同的卡上
2. 需要在Model(比如yolov5)里处理下: 如果用户设置了RuntimeOption的device_id却没有设置预处理，这个时候预处理默认与runtime保持一致  。 有些用户只知道runtime的设置，不知道预处理也要设置，所以不特殊设置预处理就默认与runtime保持一致