@heliqi 2022-11-14
> > 2. 需要在Model(比如yolov5)里处理下: 如果用户设置了RuntimeOption的device_id却没有设置预处理，这个时候预处理默认与runtime保持一致 。 有些用户只知道runtime的设置，不知道预处理也要设置，所以不特殊设置预处理就默认与runtime保持一致
> 
> @heliqi 这个逻辑不好实现呢， 因为PaddleClasPreprocessor::UseGpu(int gpu_id=0)函数来设置device id的，Preprocessor默认是不用GPU的。 只能通过传入的gpu id来设置。 或者不给这个gpu id设置默认值，让用户显式指定。

这个逻辑不要写在预处理里，预处理只管设置和处理，相关逻辑在我们Model那一层判断。 
比如不要写PaddleClasPreprocessor， 而是写在PaddleClasModel的构造函数。